split out the device model checking into NormalizedSpecific and more general code in prep for the RawData processing
